Fixed chunk-player up to a point where playing multiple chunks is possible
The problem with the previous approach is shared too much elements between
chunk branches in the pipeline.

In my first approach, each chunk branch in the pipeline had a filesrc
and matroskademux element and the rest of the elements were in the root
branch. This causes a lot of trouble.

The fix (looks like it so far) is to give each branch it's own h264parse
and avdec_h264 elements and then hook it up to the root branch.