Fixed issue where incorrect durations were written to recorded files
The problem was caused by dropping the EOS event too soon. EOS is send
onto the muxer's sink pad (left side), and was catched on the muxer's src
pad (right side). The correct approach is to send it on the muxer's
sink pad, but catch it on the filesink's sink pad instead. Then drop the
event to prevent the pipeline from going into the NULL state.

Now the durations are correctly writting to the files.